Give monkeys a small head start after they steal an item
This is adapted from a Fourk feature which largely does the same thing:
give monkeys a one-time speed boost upon stealing something that they
can use to get a head start running away from the player. The motivation
is because currently a monkey will steal something, end its turn, and
then flee away at the same speed as an un-hasted player, giving the
player some free hit opportunities; giving them a head start makes them
play more interestingly. As Fourk's commit notes, the fleeing AI in
NetHack is good enough for a monkey to give a player a decent chase, but
not good enough that the player can't hunt them down.

However, I found some problems with Fourk's implementation, so this
doesn't behave quite the same way. I didn't want to savebreak by
introducing a 'mnitro' flag to the monst struct if possible, preferring
to do it by giving the monkey bonus movement points instead. Also, I saw
an issue with mnitro in Fourk in that the monkey still ends its turn
after receiving it, and gets the speed boost at the start of its next
turn (which is what happened in my early experiments with giving the
monkey bonus movement points). In order to make the monkey's extra speed
matter, I needed to move a calculation around in movemon(). This
shouldn't cause any problems, though it is possible.

Also differing from Fourk is that the monkey gets a bonus of 12 movement
points, doubled if sped up, halved if slowed.  (In Fourk, it gets
24 even at unadjusted speed.)