Fix: cursed armor weld message could print for monsters
I haven't been able to reproduce this since first getting it, but that
initial case definitely did not involve any player gear becoming cursed.
This prevents curse() from printing weld messages unless the item being
cursed is in the hero's inventory (since cursed armor weld messages are
not implemented for in-sight monsters, this makes it not the same as
how weapons work, but I doubt anyone cares).