Generate 1/40 of non-artifact random items as pre-eroded
This is from SliceHack, though heavily modified, and is a bit of a
counterpart to the feature in which some items generate naturally
erodeproof. The SliceHack implementation had a bunch of complex rules
and made certain items (like tins) rustable which aren't eligible for
rust; this implementation tries to achieve the same goal (adding
atmosphere by not having all the random items lying around in the
dungeon looking brand-new) in a more elegant manner.

I'd rather this logic have some exemption for certain monster starting
inventories (for instance, a captain probably wouldn't go around with a
corroded saber if his underlings have un-rusted swords), but that
largely needs to happen on a case-by-case basis in the code, which I
don't think is worth it at this time. Instead, what I have done is
prevented lawful minions, prince monsters, and lord monsters from
receiving eroded gear, which covers a fair amount of cases.

Note: I specifically did not subtract the erosion of a generated weapon
when deciding whether to assign it a random macho name. So a very rusty
+2 spear can still be named Spear of Destiny, or whatever. This is so
that the player still gets a heads-up that the weapon is decently
enchanted.

Strictly speaking, the chance of an item getting eroded isn't 1/40, it's
39/1600 since the erosionproof check comes first, but close enough.