Make lieutenants M2_LORD and captains M2_PRINCE
The main reason for this is so that their gear will never be eroded by
the changes from the previous commit, and additionally captains' weapons
will always be at least +1.

The only significant side effect is that they get multishot bonuses if
they happen to be using ranged weapons. Lieutenants will get +1 to
multishot, captains +2. Makes sense to me, they're senior and have more
training.