Unify piercer-falling logic; replace AC check for evading
The intended scope for this commit was just to replace the AC-based
check with a Dex-based one, because a) the defender succeeding on this
results in an *evasion*, not a glancing blow or other thing that heavy,
enchanted armor might help with, and b) while a AC check disadvantaged
by 3 is all well and good in D&D where it comes from, NetHack's AC scale
is broken enough that an AC of 7 or below would make a piercer miss more
often than it hit, and an AC of -3 or below would make it always miss,
making them practically irrelevant. I decided on the chance of evasion
being Dex% (or monster speed% because monsters don't have Dex) - this
means that piercers will hit most of the time, but slightly less often
as the game progresses.

However, when I started implementing that, I noticed that I had never
implemented other piercer changes to the uhitm combat case where the
player polymorphed into a piercer is falling on a monster. Specifically,
the glass-piercer-breaking-glass-helm case didn't exist there, nor did
the behavior in which a hard helm only reduces 50% of damage instead of
all of it, nor did the increased base damage. (For some reason, there
doesn't appear to be an mhitm piercer-falling case.)

I really didn't want to have to reimplement all the omitted behavior in
the new uhitm cases, so I decided to refactor (again) and merge the
mhitu and uhitm codepaths into a new function, piercer_hit().