Let monsters use ranged combat in melee as a fallback option
This is from EvilHack which got it from GruntHack. The goal of this is
to avoid the common situation where, for instance, a gnome is shooting
crossbow bolts at you, then you move into melee range, and it starts
ineffectively whacking you with the crossbow even though it could
continue firing bolts.

This required a light refactor to the thrwmu() function in order to
avoid a bug present in EvilHack; if thrwmu returned void, there was no
way for mattacku to discern whether the monster had actually taken some
ranged combat action or not. Thus, in EvilHack currently the gnome will
fire its crossbow and then hit you with it, getting two attacks in one
action. By enhancing thrwmu this way, mattacku can abort from the
regular melee attack if the monster does something.

Notably, there is no logic here in which the attacking monster will
decide whether it will get more damage out of using its ranged weapon or
making a regular melee attack; it will always use the ranged weapon,
which might be relevant in cases where a monster has a high damage roll
to its AT_WEAP melee attack and mediocre ranged ammo. However, the code
doesn't currently make it easy for the AI to make estimates like that,
and it'd be very complex by nature, having to come up with expected
values for the amount of multishot, weapon damage rolls, and such.