Buff quasits, making them more distinct
Combination of two commits from Fourk. Motivation is much the same:
quasits are not a particularly interesting monster as-is; they are
unique in that they are the only monster with Dexterity poison, but this
is largely meaningless after the hero acquires resistance.

Now they generate in groups, sometimes generate invisible, can see
invisible, are faster, and can deal more Dexterity damage. To compensate
I bumped up their difficulty to 9. It doesn't solve the issue of them
becoming not a huge threat after poison resistance, but the hope is that
they feel less like popcorn this way.