Adjust backstab formula: apply a cap and give a bonus based on skill
The idea for this came from a Fourk commit, but the actual feature isn't
really the same at all. The general reasoning is that your skill
shouldn't be an irrelevant factor in how much backstab damage you do;
you'll do better backstabbing with a weapon you're good with.

Fourk takes this to its logical extreme and makes skill the *only*
factor in backstab bonus, ranging from 5 at unskilled to 28 at expert.
The intention there was to make rogues very powerful in the midgame
while not letting them be overpowered in the endgame, but by my analysis
it overpowers them through pretty much the whole game. They'd start off
getting a d10 backstab bonus for their Basic starting weapons, and it
usually only takes up until around Minetown to reach Expert in a weapon,
after which they'd have passed through an intermediate d17 bonus and
winding up with the whopping d28, for any monster they can gain an
advantage against, for most of the game.

I like the steady increase of power that comes with scaling the bonus on
XL, and it also doesn't disrupt the incentive for rogues to overlevel
past what is necessary. So that remains the main component of the bonus
in this commit. What skill does instead is cap your effective XL for
purposes of computing the bonus, as well as grant a small flat bonus
which is not subject to the random dice roll.

The caps are pretty generous; assuming the player fights with Basic
weapons, they won't notice anything until they hit level 10, at which
point they'll presumably have the ability to have a weapon they're
Skilled in. The ultimate function of the caps is that a high-level Rogue
who chooses to melee with a weapon with a lower skill cap (such as an
artifact) won't get the full backstab bonus.