Give Rogues backstab bonus on a whole bunch of other monster statuses
This is from Fourk. The motivation is to buff Rogues a bit by giving
them more opportunities to get (and engineer situations for) backstab
damage.

I was unhappy with the messages that Fourk uses, but I was also unhappy
with just leaving it as "You strike foo from behind!" since that doesn't
indicate to the player what the monster's status is, and then I learned
that there are functions which take care of formatting a monster name
with adjectives, so that solved those problems. Most of the non-fleeing
cases now produce some variation on "You deftly strike the [status]
foo!"