Monk-specific hello/goodbye messages
From UnNetHack. I'm not entirely sure about the accuracy of
"Punardarsanaya", because the majority of results in a web search for it
are NetHack related. But it seems better than using "Namaste" again;
even though "Namaste" is evidently a valid goodbye, it's more commonly
used for greeting, and could be confusing if used otherwise.