Replace Orcish Town shopkeeper survivor with Izchak
The reasons for this are twofold: it means that Izchak will be present
in every game (not that that's particularly meaningful in gameplay terms
but it's nice to have), and it is somewhat of a nerf to the amount of
loot in Orcish Town, which I have received feedback on that I've
overbuffed it by giving it too much loot. This isn't trying to rebalance
the level's loot entirely, just gently nerfing it by taking away the
price identification of arbitrary items and removing the nine random
items that the shop stocked.

Izchak occupies the place he did in Frontier Town, which is a different
shop from where the general store shopkeeper was. I was also happy to
get rid of the large complicated code block that scattered candle shop
items around the town while not actually respecting the probabilities of
candle shop items (now all the candle shop items are actually in the
shop).