Tweak a few ring-in-sink messages; polymorph control resets the sink
From UnNetHack. The only divergence between UnNetHack and here is that
when the sink transforms into another sink via dropping in polymorph
control, it does not magically create a new ring buried under the sink,
because the existence of a ring under the sink is not tracked in the
sink's looted flags in xNetHack. (The dropped-in ring will likely be
buried and can reappear with subsequent kicking, like all rings.)