Fix: dwarvish and elven ring mails gave full MC3 protection
This was a mistake; I went back in the git history and found that these
were set to MC3 when the items were first created out of removing
material-specific mithril-coats. No acknowledgement or justification was
given for why it was like this. Put them back to a base MC of 1 like
most other armors. (Note that mithril body armor always grants at least
MC2, from a separate feature.)