Fix: impossibles when backup weapon got cursed
The check for whether cursed_gear_welds should be triggered merely
checked whether owornmask was set at all. However, not all values of
owornmask actually indicate an item that's physically equipped; in
particular, W_QUIVER and W_ARTI, and W_SWAPWEP assuming the player isn't
twoweaponing. The wornmask_to_bodypart code had no result to return for
these slots, and it triggered a slew of 3 impossibles whenever some
object with one of those masks set became cursed.

There are two fixes here: to specifically check only the parts of
wornmask that refer to physical equipment, and augment
wornmask_to_bodypart to account for W_SWAPWEP which is associated with
the hands.