Fix: material bonuses in dmgval applied to more than weapons
In fact they applied to everything (or at least, any object with the
proper material which had the proper oc_dir under the hood). But they
are pretty clearly intended only for weapons ("glass is sharp" shouldn't
affect the damage when you're punching an enemy and happen to be wearing
a glass ring, for instance). So, make them only affect damage on weapons
and weapon-tools.