Fix: demon lords weren't getting their appearance messages
I don't know when exactly this happened. Perhaps at the 3.7 merge, when
the old system of `com_pager(QT_DLORD_APPEARS + foo)` was thrown out and
replaced with a string reference into quest.lua. Anyway, I believe the
error was thinking is_ndemon() meant 'is named demon', which is the
opposite of what it means. Replace it with the correct dprince and dlord
checks.