Give piranha an extra bite attack and increased speed
This is to make them ever so slightly less of a bland popcorn monster; a
piranha did not really present any threat to the hero in the places
where it was encountered. In popular media, they are pretty fast fish
that bite a lot, so a speed increase as well as an additional bite
seemed reasonable.

The extra bite is from Fourk, but not the speed increase. I declined to
take the other piranha change from Fourk that has them generate in large
groups rather than small; since they don't randomly generate, that would
only affect special levels and I'm pretty sure that group generation
doesn't even happen in special levels.