Fix: missing filled flags in various levels
This is, I think, a legacy of the filled/prefilled unification. The
default for filled on regions now being 0 meant that regions that had
previously had no need for any fill declaration at all (regions'
prefilled defaulted to 0 before this, the effect being to fill them) now
failed to get filled.

The rule of thumb in xNetHack now is that all des.regions with a type
for which filled is meaningful (e.g. special rooms) should declare the
fill status. I added it to a bunch of temples even though this doesn't
really seem to affect anything there (the priest and altar come with the
altar definition). I assigned temples filled=1 and filled=2 loosely on
the basis of there ever being some other generation that would put other
furniture or items in a temple, but it should not affect anything right
now.

Cases fixed where non-temple regions weren't getting filled:
- Barracks, a graveyard, and shops in Tou-goal
- The beehive in the Wizard's Tower