Fix: arrays declared with variable size
One inherited from the getobj patch, the other due to the gas cloud
refactor. Thanks to amateurhour for finding them.

The one in getobj was a text buffer so that's easy enough to make BUFSZ,
but the one in create_gas_cloud was declaring the array to be the size
of an argument to the function. To avoid that, I defined a maximum gas
cloud size of 150, which is larger than any current clouds should be.