Fix: kicks with bone boots still passed through ghosts
Bone is permitted to make contact with ghosts, so it didn't make any
sense that trying to kick them with bone boots failed. Initially I went
about implementing this by making all bone items get a +1 damage bonus
against incorporeal monsters, but then realized a couple problems: 1)
erosion could negate this bonus, so a kick to a ghost with thoroughly
burnt bone boots could still wind up passing through it, and 2) if you
bear hug a ghost and your torso armor comes in contact with it, there's
no real damage happening since they don't *hate* bone, just are able to
be touched by it. Thus it should only apply to kicks.

I removed the case where the only damage dealt to an incorporeal monster
would come from material hatred. Thus, as a side effect of this fix,
kicking a shade with silver boots deals the damage behind the kick in
addition to the silver damage.