Make artifactless conduct slightly more resilient
It occurred to me that if the hero could somehow engineer a scenario
where their quest artifact goes flying towards their quest leader, the
leader would give them back the artifact and this would not break
artifactless conduct. Though the in-game description is "never touch an
artifact", the conduct really means "never have an artifact in your
inventory, among a few other corner cases", so this check will break the
conduct when an artifact enters the inventory.