Make it easier on the player when kicking unknown secret doors/corridors
From Fourk. It did make some flavor sense that walls that were actually
a secret passage of some sort would mostly behave like a solid wall when
kicked, but this made it annoying for players trying to search their way
out of a room or down a corridor by kicking, because they'd most likely
just injure their legs and be unable to kick anymore for a while.

This makes it so that kicking a "wall" that is actually a secret of some
sort will never injure the player and will always produce an unambiguous
message that there's something there. This makes it a valid strategy to
kick every wall suspected of hiding something, because a single kick
will give a yes/no on whether it's a regular wall or not. It's still not
a *great* strategy, because the real walls will still hurt and wound
legs like normal.