Generate grease on an erodable object every once in a great while
This is from Fourk. Since things can generate randomly eroded or
erodeproof, why not greased as well? Won't change much in the way of
game balance, as a random greased item is less good than a random
erodeproof item, and it will be more common to find a can of grease than
to find a greased object. (Though the grease check is independent of
generating eroded/erodeproof).

Why is this a 1/23263 chance? Because that's what it is in Fourk, and I
don't have any better number in particular.