Add race-based skill cap adjustments for all nonhuman races
This is based on Fourk. The idea is that races are associated strongly
enough with certain weapons or types of magic that any member of that
race would have an aptitude for it.

Fourk dials this up pretty far, granting extra skill ranks above a
role's normal skill cap, in some cases going as high as Master. When I
calculated out what exactly that would end up being for various
race-role combinations, I found it detrimental to the balance of the
role's intended skill caps, giving them too many high skills and too
many later-game weapon options. This was probably Fourk's intent, to
give more weapon options based on race, but I am not comfortable with
that.

Instead, this implementation is fairly muted: it guarantees a certain
level of skill (Basic or Skilled) for all characters of a certain race
with certain affiliated skills.

List:
  elves get Basic enchantment spells (this is new, not in Fourk)
  dwarves get Skilled pick-axe
  orcs get Skilled scimitar
  gnomes get Basic crossbow and club