Fix: rock-breaking for flint made checks in strange order
Cavemen get more flint out of breaking rocks as a role bonus than any
character using a touchstone to do it gets. But the touchstone check
came first, so a caveman breaking rocks with a touchstone would fail to
get as much flint as a caveman breaking rocks on other rocks. Fix this
by putting the highest bonus first.