Various flint stone adjustments
Flint (and hard gems) break 50% less often. From Fourk; I'm not sure of
the original reasoning for doing this, but my reasoning is that flint
being unenchantable is harder to prevent from mulching in the later game
than other types of ammo. (You can bless it, which sets it to +2, and
that's about it.)

Hobbits that generate with a sling also generate flint (or,
occasionally, rocks) to use with it. This is both to make it slightly
more available to sling users (and to non-sling users who kill the
hobbit and get the sling), and to give sling-generating hobbits
something to use. Previously they only got to use their sling if they
happened across a gem.

Cavepeople now start the game with more flint and no rocks. I ran the
numbers and found that for cavemen banging rocks on other rocks, the
expected number of flint produced by each rock is 1. Since flint is
all-around superior to rocks (lighter, more damage), it was optimal for
a starting caveperson to break their rocks for flint one by one, which
is boring. This doesn't give quite as much additional flint as there
used to be in starting rocks, because giving them that much starting
flint seemed like way too much.

Flint now weighs 2 instead of 1. This is mostly for flavor: ideally, a
stone will weigh more than a dart or arrow; but also for balance, to
adjust them to be a little bit worse relative to other ammo types. And
this is still far lower than the usability-destroying 10 they weigh in
vanilla NetHack.