Sling buffs: strength bonus, potential to instakill H
This is a blend of the "Minor Sling Upgrade Patch" on bilious and some
of Fourk's sling buffs.

First and foremost, the melee damage bonus for strength is applied to
hits made with slung stones, because their damage comes primarily from
the ability to hurl them at high velocity. (Darts and thrown daggers,
spears, and other directly thrown weapons already enjoyed this bonus; it
was not previously given to any ammo fired from a weapon, and still
doesn't for arrows and bolts.) Fourk cuts this bonus by more than half
because of its extremely high multishot cap, which I am not porting, so
I'm not making any cuts to the bonus.

The other part of this is that a Skilled or Expert slinger can instantly
kill any H monster on a critical hit (including the Cyclops, but
practically not Surtur because Valkyries can't get that high in sling
skill). The Minor Sling Upgrade patch made this a 100% guaranteed
instakill on any hit with a sling stone ever, which is clearly
unbalanced. This implementation makes it require a critical hit, and
additionally only works on 50% of critical hits if merely Skilled and
not Expert.