Create statues when footrice generate during level creation
From Fourk. This is to make the level a bit more interesting when
trices generate and/or give the player an early warning that there's one
around. The statue ends up getting placed within a 11x11 square centered
on the footrice on walkable terrain. (This can potentially do things
like put a statue in an inaccessible vault if the footrice spawned
close enough to it; there's no pathfinding check to see if the square is
actually reachable from the monster.)

Changes from Fourk's implementation:
- I cut the number of generated statues significantly; Fourk will almost
  always produce 6 in an open area, less with more walls around. This
  tries to create 1, with a recursive 1/4 chance for more.
- Since chickatrices generate in small groups, each individual
  chickatrice only gets a 1/3 chance of attempting to generate statues.
  In Fourk a crowd of four chickatrices could generate more statues than
  on Medusa's level.
- I blocked statue creation if the footrice is spawning in a special
  room, so that zoos and cockatrice nests don't also generate (even
  more) statues.
- Fourk uses an 11x7 rectangle for potential statue locations for some
  reason. I don't like basing gameplay on the fact that terminal
  characters have a aspect ratio higher than they're tall.