New themed room: Pool challenge from /dev/null/nethack
This is a lookalike room to the pool challenge, containing six "pockets"
(holes; the original challenge had pits, but there isn't a simple way to
make the pits pre-mapped, whereas holes are always pre-mapped) and nine
"balls" (boulders). Unlike in devnull, the boulders are numbered 1-9,
because that's a cool detail.

This isn't a reimplementation of the challenge - no bright magenta cue
ball, cue stick, or mechanism for knocking the boulders into each other;
it's just a homage.