Hardware stores stock potions of oil, and fewer touchstones
The potions of oil in hardware stores are from Fourk. While I was in
that area of the code, I reduced the 7% chance of touchstones, which had
been decided earlier was too high (particularly notable in larger
stores where there were three to five of them; the player typically only
needs one).