Make yellow lights and fire elementals emit radius 2 light
Been planning this for a while. The idea is that light-emitting monsters
that are "brighter" should emit more light, like yellow lights are
brighter than black lights and fire elementals (so I imagine them as)
burn brighter than mere vortices or spheres.

In vanilla, this apparently existed at some point but was then cut back
to radius 1 light for all monsters for 'performance reasons'. Consider
this a de facto drop of support for all systems that would struggle
performance-wise to run the computations associated with monster light
radii of 2.