Support holiday-themed fruits for many more holidays
This started out by porting Fourk's feature for the same thing, which is
primarily US holidays and foods from the advent calendar patch, and I
decided I wanted to go a bit more multicultural.

In particular, this tracks major Islamic and Jewish holidays, which
necessitated adding a whole bunch of code for translating Gregorian
calendar dates into Hebrew and Hijri dates. Also, Dia de Muertos is
tracked, but that uses a straightforward Gregorian date.

The holiday system that this commit creates is capable of correctly
finding multiple holidays that fall on the same day and having an
arbitrary number of associated food names with each holiday. Not all
holidays are associated with themed foods - for instance, fast holidays
such as Yom Kippur and Ramadan don't have any one particular food that
is commonly eaten at that time, so for now such holidays aren't actually
used anywhere.

I also added encyclopedia entries for the foods that are possibly less
well known. And a couple special cases to the pluralization code for
Spanish and Hebrew foods that pluralize irregularly.