Extend the target range for stinking clouds
Another Fourk feature, but not quite identical. I wanted to port its
behavior where the beatitude of the scroll gives a greater or lesser
radius for targeting the cloud, but unfortunately in 3.6 the radius
check was moved into a function I can't add arguments to because it's a
getpos_sethilite callback and I didn't want to add a global variable for
this one case.

But I still wanted to increase the range, so I set it to what Fourk uses
for its uncursed scroll, which is 50 (in practice giving it up to 7
spaces orthogonally and up to 1 perpendicular to it).