Make gas clouds block line of sight, and print a dissipation message
This has always been something that bothers me. In what system of
physics are you obscured from seeing the area within a cloud but can see
the area behind the cloud with perfect clarity?

From Fourk, with modifications (again). I'm not sure why Fourk doesn't
use the existing does_block function to compute whether a point being
losing its gas cloud should remain blocked or not. I'm also not sure why
Fourk doesn't add a case for gas clouds in does_block.

The other new addition here is that when you see or experience a cloud
vanish, it always prints a message. This supersedes the old cases that
printed a message only in response to you praying to get rid of region
trouble. I think that this is worth it to get rid of the oddity of gas
clouds vanishing completely silently, even if the unfortunate side
effect is to add more message spam to the Plane of Fire.

I declined to port the Fourk behavior where hezrou are able to see
through stinking clouds by virtue of having astral vision - I wanted to
add neither a bunch of special exemptions for player and monster hezrous
in the vision code, nor allow players to get free astral vision by
polymorphing into a hezrou.