Make hallucination block petrification
Another Fourk feature, implementing one of the greatest ever YANIs for a
pun based mechanic - you don't turn to stone because you're already
stoned.

The major implications of this change are that you can employ
hallucination to wield a cockatrice corpse without gloves, or perhaps as
a safeguard against accidentally falling into a pit with it; and that
you can use it to protect yourself completely from hissing attacks. I
don't see a problem with either; it gives hallucination some additional
valid uses without being so obviously good that it would clearly be
used by everyone.

You will get stoned if the hallucination expires while wielding a
cockatrice corpse, and in the edge case of when you have hallucination
active but suppressed by wielding Grayswandir and you change to wielding
a cockatrice corpse instead.