Add more spears in monster starting inventory
Yet another Fourk feature; the intention is to make spears a bit more
common for spear-users. Vanilla already added the possibility of dwarves
getting a spear, so the only monsters left without any chance of spears
were orcs.

The other part of this feature, also from Fourk, is increasing the
quantity of spears any monster will receive when it generates with
spears, with an expected value of about 2 spears (but biased towards 1).