Credit the player with monster trap deaths on player's turn
If thitm killed a monster under any circumstances, the player was not
credited, regardless of the player setting the trap or whether the
player forced the monster onto that trap, e.g. with a staggering blow.
This doesn't do anything about the player-set-trap case (in part because
it seems unfair to pacifists that any stray pits they might leave around
could randomly kill a wandering monster and break their conduct), but it
makes it so that all trap deaths on the player's turn count as player
kills.

This is largely in preparation for an upcoming feature in which the
player can very clearly directly kill a monster with trap damage and
needed to be credited, but it was good to fix cases like jousting and
martial arts anyway.