Polymorphs into most snaky monsters don't break armor
This is another Fourk feature, aimed to address the oddity of busting
out of one's armor if turning into a large snake like a naga or python
which are not actually thicker than a typical Medium-sized creature's
torso.

Unlike Fourk I placed the size cutoff at Huge, meaning that turning into
things like nagas will not break armor (since they really only ought to
be about as big around as a human's body). I also went and implemented
it in such a way as to be symmetric for monsters.

I found a minor vanilla bug in the process of implementing this, but
left it in place since I got the fix into vanilla already.