Chaotics do not get alignment penalties for angering peacefuls
Another Fourk feature, which makes sense to me - why would a chaotic god
take any particular offense to you randomly getting in fights with
civilians?

This probably has little effect in the game right now, since an
alignment penalty of -1 is practically meaningless anyway, but it might
be a useful marker for if any alignment overhaul comes down the line at
some point.