Pre-populate teleport destination prompt with travel destination
Another Fourk feature, though this one was a FIQ commit so it may have
originated from FIQHack. The main quality of life this provides is for
when you are traveling across the level with controlled teleportitis;
you no longer have to suffer the annoyance of moving the cursor to your
destination again.

The only place I can see this backfiring is if someone were trying to
travel across the edge of a teleport region, at which point trying to
teleport there would Sorry them and send them somewhere random. But that
seems unlikely to happen much.