Adjust the formula for the number of summoned nasties
Another Fourk change. This formula change also introduces a related
Fourk feature, which is that summon nasties' monster count should scale
based off the caster's level, not the player's. (Why should Rodney and
liches be kinder to a low-level pacifist player?)

The goal of the formula change is to make the summon nasties range a bit
less swingy, where level 30 could produce as many as 10 or as few as 1
monster.

The new formula makes the maximum number of possible monsters scale up
more slowly with respect to level, but increases the minimum to 3
instead of 1. (Fourk sets the minimum at 4 which I felt was a bit too
high.) A level 30 Wizard will generate about the same amount of nasties
as he did previously; mid-level casters will generate slightly more.