Add L's Wounds patch: healers can see damage on monsters
This implementation is largely based on Fourk's, but that seems pretty
true to the original patch. As usual, I did reimplement it with
modifications rather than just copying, to get the logic in mon_wounds
better organized, to pick a slightly different set of adverbs, and to
extend the wounds strings produced by hallucination.

This produces wound messages when a healer damages a monster with: a) a
melee attack, b) a ranged weapon, c) a beam spell, d) a ray spell.
Additionally, I adopted a feature from EvilHack in which a message is
also produced when a pet is hurt in mvm combat. I felt like this might
be too much message spam, but evidently Fourk, UnNetHack, and EvilHack
have had this feature for a while and it hasn't been a huge issue there.
Additionally, in order to avoid printing a wounds message when doing a
non-damaging attack to a monster (e.g. zapping it with sleep), it only
prints the message when the adverb would change. Thus, if you are
wailing on a seriously wounded monster, it will not print wounds
messages for hits that keep it seriously wounded, only when it becomes
badly wounded. (Also, if it silently regenerates HP to be classified as
moderately wounded, it will then print a seriously wounded message upon
taking further damage, so this isn't perfect.)

More interestingly, healers also get to see the wounded state of any
monster in sight via farlook, which is less accurate but more powerful
than a stethoscope, with no melee-range requirement and no limit on the
number of monsters viewed per turn.