Fix: use the correct math for deciding to make a door secret
Thanks to paxed for pointing this out, that dodoor() was actually making
secret doors far less often than it should have been, due to my dumb
math failure of assuming that squaring the numerator and denominator of
a fraction doesn't change its value. The correct formula is presented in
the code comments.