Obj materials: don't allow non-burnable containers to burn up
This is an oversight in the obj materials system; vanilla assumes all
containers (except ice boxes, which get a special exception) are capable
of burning up instantly. This is not true under object materials, which
allows things like iron and stone boxes. Add a check for this. Thanks to
K2 (or whoever reported the bug in EvilHack) for pointing this out.