Decrease damage done by digging out of engulfer
This was previously cut specifically for Juiblex, but I wasn't fully
happy with it: it still seemed too cheap and easy to bust out of
engulfers. So, I have applied that logic (cut HP in half) to all
engulfers, and decreased it further for amorphous engulfers (which
includes Juiblex). It's still a valid strategy to dig out of them to cut
a big chunk of their HP off, but that can't be the entire strategy.