Give "You begin bashing" message for unskilled weapons/tools
This is from FIQHack. I believe the goal is to help remove the perennial
problem of going around for thousands of turns hitting monsters with a
pickaxe because the player forgot to switch back to their real weapon
after doing some digging.

It will be a slight annoyance (a bit extra message spam) for players who
are switching to an unskilled weapon to train it. But the benefit
outweighs that annoyance in my opinion.

Also, it needs to check g.moves > 1 to avoid giving a bashing message
with the default wielding weapon; there is a case in which wielding a
(different) unskilled weapon on turn 1 will not give the bashing
message, but that's a pretty marginal edge case.