Make mold grow on corpses substantially less often
The number we're targeting here is the overall chance a corpse will go
moldy (calculated via taking a binomial distribution of 1/MOLDY_CHANCE
over 200 trials and examining the probability of >= 1 successes). I cut
this from roughly 50% to roughly 5%.

The reason is simply because making half of corpses go moldy was way too
much: going through and slaying a leprechaun hall or other room full of
monsters would wind up with a room full of fungi, often popping up
before the room was even cleared. Returning to a cleared level was
similar. Since F are a source of intrinsics and permafood, this likely
made that rather easier to get than I intended. So the going moldy event
should overall be rarer than it already was.