Refactor: dump_container() to unify item-spilling code
This is mostly in preparation for an upcoming feature, but I saw the
opportunity to apply it to some other code paths as well (such as items
spilling out of a stone-to-fleshed statue turning into meatballs), thus
making this a code-unification commit. The code is healthier unified: it
prevents weirdness with items spilling in shops and possible (or just
latent) exploits with never-rotting corpses spilled out of an ice box.